Organize the folders and project files
