+++ src/components/AnimeCard/AnimeCard.style.css
