+++ src/components/AnimeCard/index.js
