+++ src/components/AnimeDetails/AnimeDetails.style.css
